+++ posit/posit
@@ -56,6 +56,6 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////
 /// math functions
-#include "math_functions.hpp""
+#include "math_functions.hpp"
 
 #endif
