+++ tests/posit/postfix.cpp
@@ -1,6 +1,6 @@
 // postfix.cpp functional tests for postfix operators
 //
-// Copyright (C) 2017-2018 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2019 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
