+++ tests/posit/specialized/256bit_posit.cpp
@@ -13,7 +13,7 @@
 #define POSIT_THROW_ARITHMETIC_EXCEPTION 0
 #include <posit>
 #include "../../test_helpers.hpp"
-#include "../../posit_test_helpers.hpp"
+#include "../../posit_test_randoms.hpp"
 
 /*
 Standard posits with nbits = 256 have 5 exponent bits.
@@ -49,10 +49,10 @@ try {
 #if STRESS_TESTING
 	cout << "Arithmetic tests " << RND_TEST_CASES << " randoms each" << endl;
 	cout << "Without an arithmetic reference, test failures can be ignored" << endl;
-	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<128, 4>(tag, bReportIndividualTestCases, OPCODE_ADD, RND_TEST_CASES), tag, "addition      ");
-	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<128, 4>(tag, bReportIndividualTestCases, OPCODE_SUB, RND_TEST_CASES), tag, "subtraction   ");
-	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<128, 4>(tag, bReportIndividualTestCases, OPCODE_MUL, RND_TEST_CASES), tag, "multiplication");
-	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<128, 4>(tag, bReportIndividualTestCases, OPCODE_DIV, RND_TEST_CASES), tag, "division      ");
+	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<nbits, es>(tag, bReportIndividualTestCases, OPCODE_ADD, RND_TEST_CASES), tag, "addition      ");
+	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<nbits, es>(tag, bReportIndividualTestCases, OPCODE_SUB, RND_TEST_CASES), tag, "subtraction   ");
+	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<nbits, es>(tag, bReportIndividualTestCases, OPCODE_MUL, RND_TEST_CASES), tag, "multiplication");
+	nrOfFailedTestCases += ReportTestResult(ValidateBinaryOperatorThroughRandoms<nbits, es>(tag, bReportIndividualTestCases, OPCODE_DIV, RND_TEST_CASES), tag, "division      ");
 #endif
 	nrOfFailedTestCases = 0;
 	return (nrOfFailedTestCases > 0 ? EXIT_FAILURE : EXIT_SUCCESS);
